requirements.txt
pandas==1.5.2
biopython==1.80
requests==2.28.2